% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerOrdinal.R
\name{LearnerOrdinal}
\alias{LearnerOrdinal}
\title{Ordinal Regression Learner}
\format{
\link{R6Class} object inheriting from \link{Learner}.
}
\description{
This Learner specializes \link{Learner} for ordinal regression tasks.
}
\section{Usage}{

See \link{Learner}.
}

\examples{
library(mlr3)
ids = mlr_learners$keys("^ordinal")
lrns = mlr_learners$mget(ids)
names(lrns)

# get a specific learner from mlr_learners:
lrn = mlr_learners$get("ordinal.clm")
print(lrn)
}
\concept{Learner}
