% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeasureOrdinal.R
\name{MeasureOrdinal}
\alias{MeasureOrdinal}
\title{Ordinal Regression Measure}
\format{\link{R6Class} object inheriting from \link{Measure}.}
\description{
This task specializes \link{Measure} for ordinal regression problems.
}
\section{Usage}{

See \link{Measure}.
}

\section{Details}{

\code{$task_type} is \code{"ordinal"}.
}

