% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{mlr3ordinal-package}
\alias{mlr3ordinal}
\alias{mlr3ordinal-package}
\title{mlr3ordinal: Ordinal regression for mlr3}
\description{
Provides extensions for ordinal regression for mlr3.
  This includes a specialized task type, models, and measures.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mlr-org/mlr3ordinal}
  \item Report bugs at \url{https://github.com/mlr-org/mlr3ordinal/issues}
}

}
\author{
\strong{Maintainer}: Stefan Coors \email{mail@stefancoors.de} (\href{https://orcid.org/0000-0002-7465-2146}{ORCID})

Other contributors:
\itemize{
  \item Michel Lang \email{michellang@gmail.com} [contributor]
}

}
