% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_learners_ordinal_rpart.R
\docType{data}
\name{mlr_learners_ordinal.rpart}
\alias{mlr_learners_ordinal.rpart}
\alias{LearnerOrdinalRpart}
\title{Ordinal Tree Learner}
\format{\link[R6:R6Class]{R6::R6Class} inheriting from \link{LearnerOrdinal}.}
\usage{
LearnerOrdinalRpart
}
\description{
A learner for a regression tree implemented in \code{\link[rpart:rpart]{rpart::rpart()}}.
Thresholds on predictions are optimized by \code{\link[GenSA:GenSA]{GenSA::GenSA()}}.
}
\keyword{datasets}
