% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_measures_mre.R
\docType{data}
\name{mlr_measures_mre}
\alias{mlr_measures_mre}
\alias{MeasureOrdinalMRE}
\title{Mean Rank Error Measure}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link{MeasureOrdinal}.}
\usage{
MeasureOrdinalMRE
}
\description{
Calls \link[Metrics:ce]{Metrics::ce}.
}
\keyword{datasets}
