% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_measures_mze.R
\docType{data}
\name{mlr_measures_mze}
\alias{mlr_measures_mze}
\alias{MeasureOrdinalMZE}
\title{Mean Zero-One Error Measure}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link{MeasureOrdinal}.}
\usage{
MeasureOrdinalMZE
}
\description{
Calls \link[Metrics:ce]{Metrics::ce}.
}
\keyword{datasets}
