% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeasureOrdinalACC.R
\docType{data}
\name{mlr_measures_ordinal_acc}
\alias{mlr_measures_ordinal_acc}
\alias{MeasureOrdinalACC}
\title{Mean Accuracy Measure}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link{MeasureOrdinal}.}
\usage{
MeasureOrdinalACC
}
\description{
Calls \link[Metrics:ce]{Metrics::ce} for calculating the classification error. Resulting ACC = 1 - CE.
}
\keyword{datasets}
