% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PipeOpConvertOrdinalTask.R
\docType{data}
\name{mlr_pipeop_convertordinaltask}
\alias{mlr_pipeop_convertordinaltask}
\alias{PipeOpConvertOrdinalTask}
\title{PipeOpConvertOrdinalTask}
\format{\code{\link{R6Class}} object inheriting from \code{\link{PipeOpTaskPreproc}}.}
\usage{
PipeOpConvertOrdinalTask
}
\description{
Converts an ordinal task to a regression one.
}
\seealso{
Other PipeOps: \code{\link{mlr_pipeop_ordinalthresholds}}
}
\concept{PipeOps}
\keyword{datasets}
