% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PipeOpOrdinalRegr.R
\name{mlr_pipeop_ordinalregr}
\alias{mlr_pipeop_ordinalregr}
\alias{PipeOpOrdinalRegr}
\title{PipeOpOrdinalRegr}
\format{\link{R6Class} PipeOpOrdinalRegr}
\description{
This PipeOp works for any regression learner.
The idea is to consider the whole task as a \code{\link[mlr3:TaskRegr]{Regression Task}}.
Hence, a numeric response vector is predicted for which crossvalidated rank thresholds are found using (nloptr)\link[nloptr:nloptr]{nloptr::nloptr} to rank the predictions.
Returns a single \code{\link{PredictionOrdinal}}.
As a default, optimizes \code{\link{MeasureOrdinalCE}}.
Used for regression \code{\link{Prediction}}s.
}
\examples{
library(mlr3pipelines)
op = po("ordinalregr", 2)
}
\seealso{
Other PipeOps: 
\code{\link{mlr_pipeop_convertordinaltask}},
\code{\link{mlr_pipeop_ordinalclassif}}
}
\concept{PipeOps}
