% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PipeOpOrdinalRegression.R
\docType{data}
\name{mlr_pipeop_ordinalregression}
\alias{mlr_pipeop_ordinalregression}
\alias{PipeOpOrdinalRegression}
\title{PipeOpOrdinalRegression}
\format{\link{R6Class} PipeOpOrdinalRegression}
\usage{
PipeOpOrdinalRegression
}
\description{
This PipeOp works for any regression learner.
The idea is to predict numeric target values and optimizing ordinal class thresholds afterwards to label the predictions.
Here, optimal thresholds are searched over different \code{\link{PredictionRegr}}s.
Ordinal thresholds for each regression learner are optimized using (GenSA)\link[GenSA:GenSA]{GenSA::GenSA}.
Returns a single \code{\link{PredictionOrdinal}}.
As a default, optimizes \code{\link{MeasureOrdinalMMCE}}.
Used for regression \code{\link{Prediction}}s.
}
\examples{
op = PipeOpOrdinalRegression$new()
}
\concept{PipeOps}
\keyword{datasets}
