% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaskOrdinalWinerating.R
\name{mlr_tasks_winerating}
\alias{mlr_tasks_winerating}
\alias{load_task_winerating}
\title{Wine Rating Ordinal Task}
\format{
\link[R6:R6Class]{R6::R6Class} inheriting from \link{TaskOrdinal}.
}
\description{
An ordinal task for the \link[ordinal:wine]{ordinal::wine} data set.
}
\section{Usage}{
\preformatted{mlr_tasks$get("winerating")
}
}

