% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_ordinal_threshold.R
\name{optimize_ordinal_threshold.R}
\alias{optimize_ordinal_threshold.R}
\alias{optimize_ordinal_threshold}
\title{Optimization loop for obtaining optimal ordinal regression threshold}
\usage{
optimize_ordinal_threshold(learner, task)
}
\description{
A funtion which uses predictions of any regression learner und tunes the threshold to get ordinal ranks.
Thresholding is done by \code{\link[GenSA:GenSA]{GenSA::GenSA()}}.
}
