% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numerics_to_ordinal.R
\name{optimize_ordinal_threshold}
\alias{optimize_ordinal_threshold}
\title{Optimization loop for obtaining optimal ordinal regression threshold}
\usage{
optimize_ordinal_threshold(learner, task)
}
\description{
Optimization loop for obtaining optimal ordinal regression threshold
}
