% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PipelineOrdinalThreshold.R
\docType{data}
\name{PipelineOrdinalThreshold}
\alias{PipelineOrdinalThreshold}
\title{PipelineOrdinalThreshold}
\format{\code{\link{R6Class}} object inheriting from \code{\link[mlr3pipelines:Graph]{mlr3pipelines::Graph}}.}
\usage{
PipelineOrdinalThreshold
}
\description{
A pipeline that that takes any \code{\link[mlr3:LearnerRegr]{Regression Learner}} to be used for \code{\link{TaskOrdinal}}.

The result is a \code{\link[mlr3pipelines:Graph]{Graph}} which works as regular learner when put into \code{\link[mlr3pipelines:GraphLearner]{GraphLearner}}.
}
\keyword{datasets}
