% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictionOrdinal.R
\name{PredictionOrdinal}
\alias{PredictionOrdinal}
\title{Prediction Object for Ordinal Regression}
\format{\link{R6Class} object inheriting from \link{Prediction}.}
\description{
This object stores the predictions returned by a learner of class \link{LearnerOrdinal}.
}
\section{Usage}{

Inherits from \link{Prediction}\preformatted{# Construction
p = PredictionOrdinal$new(task, risk)

# Members
p$predict_types
p$row_ids
p$truth

# S3 methods
as.data.table(p)
}
}

\section{Arguments}{

\itemize{
\item \code{task} (\link{Task}):
Task used for prediction. Used to extract \code{row_ids} and \code{truth}.
Set to \code{NULL} to skip all argument checks during initialization.
Slots \code{p$row_ids} and \code{p$truth} need to be set manually in this case.
\item \code{response} (\code{factor()} | \code{ordered()}): Vector of predicted class labels.
\item \code{prob} (\code{matrix}):
Numeric matrix of rank probabilities with one column for each rank in \code{task$all_classes}
and one row for each observation in the test set.
}
}

\section{Details}{

\itemize{
\item \code{$new()} initializes a new object of class \link{Prediction}.
\item \code{$predict_types} (\link{character}) stores the predict types available: a subset of \code{c("response", "prob")}.
\item \code{$response} stores the predicted values.
\item \code{row_ids} stores the row IDs.
\item \code{$truth} stores the true rank vector.
\item The prediction object can be transformed to a simple \code{\link[=data.table]{data.table()}}
with \link[data.table:as.data.table]{data.table::as.data.table}.
}
}

\examples{
library(mlr3)
task = mlr_tasks$get("wine")
learner = mlr_learners$get("ordinal.clm")
e = Experiment$new(task, learner)$train()$predict()
p = e$prediction
head(as.data.table(p))
}
\concept{Prediction}
