% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numerics_to_ordinal.R
\name{RegrToOrdinal.R}
\alias{RegrToOrdinal.R}
\alias{numerics_to_ranks}
\title{Transformation of numeric response to ranks}
\usage{
numerics_to_ranks(response, task)
}
\description{
A funtion which uses predictions of any regression learner und tunes the threshold to get ordinal ranks.
Thresholding is done by \code{\link[GenSA:GenSA]{GenSA::GenSA()}}.
}
