% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaskOrdinal.R
\name{TaskOrdinal}
\alias{TaskOrdinal}
\title{Ordinal Task}
\format{
\link{R6Class} object inheriting from \link[mlr3:Task]{mlr3::Task}/\link[mlr3:TaskSupervised]{mlr3::TaskSupervised}.
}
\description{
This task specializes \link[mlr3:Task]{mlr3::Task} and \link[mlr3:TaskSupervised]{mlr3::TaskSupervised} for ordinal regression problems.
}
\section{Usage}{

Inherits from [mlr3::Task]/[mlr3::TaskSupervised].\preformatted{# Construction
t = TaskSupervised$new(id, backend, target)


[mlr3::Task]: R:mlr3::Task
[mlr3::TaskSupervised]: R:mlr3::TaskSupervised
}
}

\examples{
library(mlr3)
# ...
}
