% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeasureOrdinalMAE.R
\docType{data}
\name{mlr_measures_ordinal.mae}
\alias{mlr_measures_ordinal.mae}
\alias{MeasureOrdinalMAE}
\title{Mean Absolute Error Measure}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link{MeasureOrdinal}.}
\usage{
MeasureOrdinalMAE
}
\description{
Calls \link[Metrics:mae]{Metrics::mae}.
}
\keyword{datasets}
