% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PipeOpConvertOrdinalTask.R
\docType{data}
\name{mlr_pipeop_convertordinaltask}
\alias{mlr_pipeop_convertordinaltask}
\alias{PipeOpConvertOrdinalTask}
\title{PipeOpConvertOrdinalTask}
\format{\code{\link{R6Class}} object inheriting from \link[mlr3pipelines:PipeOpTaskPreproc]{mlr3pipelines::PipeOpTaskPreproc}.}
\usage{
PipeOpConvertOrdinalTask
}
\description{
Converts an \link{TaskOrdinal} to an \link[mlr3:TaskRegr]{mlr3::TaskRegr} or \link[mlr3:TaskClassif]{mlr3::TaskClassif} one, depending on \code{type}.
}
\seealso{
Other PipeOps: \code{\link{mlr_pipeop_ordinalthresholds}}
}
\concept{PipeOps}
\keyword{datasets}
