% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PipeOpOrdinalClassif.R
\name{mlr_pipeop_ordinalclassif}
\alias{mlr_pipeop_ordinalclassif}
\alias{PipeOpOrdinalClassif}
\title{PipeOpOrdinalClassif}
\format{\link{R6Class} PipeOpOrdinalClassif}
\description{
This PipeOp works for any classification learner.
The idea is to discard the ordinal structure of the target variable and consider the whole task as a \code{\link[mlr3:TaskClassif]{Classification Task}}.
Returns a single \code{\link{PredictionOrdinal}}.
As a default, optimizes \code{\link{MeasureOrdinalCE}}.
}
\examples{
library(mlr3pipelines)
op = po("ordinalclassif")
}
\seealso{
Other PipeOps: 
\code{\link{mlr_pipeop_convertordinaltask}},
\code{\link{mlr_pipeop_ordinalregr}}
}
\concept{PipeOps}
