% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaskOrdinalRegr.R
\docType{data}
\name{TaskOrdinalRegr}
\alias{TaskOrdinalRegr}
\title{Ordinal Regression Task}
\format{\link[R6:R6Class]{R6::R6Class} object inheriting from \link{TaskRegr}/\link{Task}/\link{TaskSupervised}.}
\description{
This task specializes \link{Task} and \link{TaskSupervised} for ordinal regression problems.
If the target column is an ordered factor, it is transformed to a numeric vector.
The original orderd target is stored in the \link{DataBackend} with \link{col_role = "target_ordinal"}
Predefined tasks are stored in \link{mlr_tasks}.

The \code{task_type} is set to \code{"regr"}.
}
\section{Construction}{
\preformatted{t = TaskOrdinalRegr$new(id, backend, target, target_ordinal)
}
\itemize{
\item \code{id} :: \code{character(1)}\cr
Name of the task.
\item \code{backend} :: (\link{DataBackend} | \code{data.frame()} | ...)\cr
Either a \link{DataBackend}, or any object which is convertible to a DataBackend with \code{as_data_backend()}.
E.g., a \code{data.frame()} will be converted to a \link{DataBackendDataTable}.
\item \code{target} :: \code{character(1)}\cr
Name of the target column.
\item \code{target_ordinal} :: \code{character(1)}\cr
Name of the original ordered target column.
}
}

\section{Fields}{

}

\section{Methods}{

}

\examples{
task = TaskOrdinalRegr$new("wine", backend = iris, target = "rating")
task$task_type
task$formula()
task$truth()

# possible properties:
mlr_reflections$task_properties$ordinalregr
}
\concept{Task}
\keyword{datasets}
